requirements.txt
altgraph==0.17.4
babel==2.17.0
packaging==24.2
pefile==2024.8.26
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
PySide6==6.8.2.1
PySide6_Addons==6.8.2.1
PySide6_Essentials==6.8.2.1
python-dotenv==1.0.1
pywin32-ctypes==0.2.3
setuptools==75.8.2
shiboken6==6.8.2.1
